# docs/swagger.yaml
definitions:
...
  api.updateLaporanMasyarakatRequest:
    properties:
      hubungan_dengan_balita:
        type: string
      id:
        type: string
      id_balita:
        type: string
      id_masyarakat:
        description: dapat null jika laporan dari admin
        type: string
      id_status_laporan:
        type: string
      nomor_hp_keluarga_balita:
        type: string
      nomor_hp_pelapor:
        type: string
      tanggal_laporan:
        description: 'Format: YYYY-MM-DD'
        type: string
    type: object
  api.updateLaporanMasyarakatResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
...
  object.Response:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing stunting data
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Stunting Web API
  version: 0.0.2
paths:
...
  /api/admin/laporan-masyarakat/update:
    put:
      consumes:
      - application/json
      description: |-
        Update existing laporan masyarakat data (Admin only)

        Updates laporan masyarakat record with new data including:
        - id_masyarakat (optional, null if admin report), id_balita, id_status_laporan
        - tanggal_laporan, hubungan_dengan_balita, nomor_hp_pelapor, nomor_hp_keluarga_balita
        - Validates balita existence, status laporan, masyarakat (if provided), and prevents duplicates
      parameters:
      - description: Updated laporan masyarakat data
        in: body
        name: laporan
        required: true
        schema:
          $ref: '#/definitions/api.updateLaporanMasyarakatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Laporan masyarakat updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.updateLaporanMasyarakatResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Laporan masyarakat not found
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - Bearer: []
      summary: Update laporan masyarakat data
      tags:
      - admin
...
swagger: "2.0"